Create README.md
"All Stock Anaylsis for Challenge Two"

Overview of "All Stock AnalysiOverview of "All Stock Analysis":
The purpose of this project; "All Stock Analysis was to compare the original coding run for twelve stocks in 2018 from NASDAQ exchange for Steve and his parents. Then Steve wanted to compare the top twelve stocks with the highest total overall total daily volume and compare the return for each annual year in which he was investing and anaylzing the best stock performance for two annual periods. The returns reflected unknown behaviors in the markets for each of the stocks. The top three company for 2017 made significant returns were DAQO New Energy Corporation ADR, (DQ), a Chinese company primarily involved in solar photovoltaic systems with a return of 199.4%. with was a high performance stock. Secondly, SolarEdge Technologies Inc.,((SEDG), an Israeli company involved in solar inverters photovoltaic arrays and other products indicated an impressive return of 184.5%. Next, a third soaring energy company with a return 129.5% (2017) was Enphase Energy Inc. (ENPH) an American company manufacturing solar panels and PV supporitng systems. ![2022-08-10 19.08.01.png](path/to/2022-08-10 19.08.01.png).(refer to diagram #1) Lastly First Solar, Inc. an American manufacturer of solar panels and PV supporting systems with the second largest comparable total daily volume came in with a mild return of 101.3%. The project was labeled by their ticker symbol in accordance with NASDAQ as energy stocks and parsed the top 12 stocks which is something they can reflex on base the annual return.
Notice from diaram #4 that the 2018 stocks plummed to mostly negative returns, with the exception of Enphase Energy Inc.(ENPH) an American company, 81.9% and SUNrun Inc.(RUN) 84.0% respectively.
Results:
 The Tables implicate the top daily volume with in each annual year on twelve stocks for 2017 and 2018 reflecting each stock's return for that given year. The overall VBA process parsed out information pertaining to twelve stocks, next their overall total daily volume and lastly the return of that stock. 2017 total return for the twelve stocks were positive except for RUN. However, in 2018 was extensive negative returns. The overall increase in total daily volume was 139,399,100 from 2017 to 2018. It required a ticker index to be established with numerous steps in the coding process.
The two years were created to compare against the original and refactored coding in VAS for two years repectively in 2017, and 2018. As you can see in the diagram #2 and #3 the screen shot the code ran faster in 2017 0.05859375 seconds vs. 2018 0.9140625 seconds. I had to continue to debug and then the excel system would crash. In turn I re-wrote the code again. The steps to coding were provided as a statement. The overall daily total volume increased by 139,399,100 for the top twelve stocks from 2017 to 2018. The overall experience was that the 2017 had a a lower daily voulume holding so that the stock returns can not be determined what the indicators could be conditions to a lucrative positive return. If additional codes were added to compare the returns reflection of why it success was consistent would be another possible consideration. 
Summary: 
 The coding process looped was established to link the worksheet. The objective was to increase the ticker volume for the volume current ticker. 
 A next point to be made are the details statements that are advantageous comparably original and refactored VBA scripts were the speed and the overall volume was less in the in the original information to code. The original stock analysis was just for the one-year 2018. So, the code was milder comparatively. The speed was also quicker on the original all stock analysis vs. All stock Analysis refactored. The disadvantage for coding the All-Stock Analysis refactored process was that it contained an extensive amount of data in which it took longer to process.